+++ projekt/Bank/Bank/Przelew.h
@@ -11,18 +11,18 @@ using namespace std;
 class Przelew {
 
 private:
-	unsigned long int id_odbiorca;
+	long long int id_odbiorca;
 	double kwota;
-	unsigned long int id_przelewu;
+	long long int id_przelewu;
 	string typ_przelewu;
 	long long int ID;
 
 public:
 
 
-	void stworz_przelew(unsigned long int id_odbiorca, double kwota);
+	void stworz_przelew(long long int id_odbiorca, double kwota);
 
-	void blik_menu(Blik*& nowy, double k, unsigned long int id_od);
+	void blik_menu(Blik*& nowy, double k, long long int id_od);
 
 	bool wyslij_przelew();
 
@@ -32,13 +32,13 @@ public:
 
 	void kod_weryfikacji();
 
-	void zapisz_przelew_w_bazie(unsigned long int id_odbiorca, string typ_przelewu);
+	void zapisz_przelew_w_bazie(long long int id_odbiorca, string typ_przelewu);
 
 	void potwierdz_przelew();
 
 	void get_informacje();
 
-	Przelew(unsigned long int id_o = 0, double kwo = 0, unsigned long int id_p = 0, string t_p = "", long long int ID = 0) : id_odbiorca(id_o), kwota(kwo), id_przelewu(id_p), typ_przelewu(t_p), ID(ID)
+	Przelew(long long int id_o = 0, double kwo = 0, long long int id_p = 0, string t_p = "", long long int ID = 0) : id_odbiorca(id_o), kwota(kwo), id_przelewu(id_p), typ_przelewu(t_p), ID(ID)
 	{
 
 	}
